requirements.txt
fastapi~=0.112.2
SQLAlchemy~=2.0.32
pydantic~=2.8.2
python-jose~=3.3.0
passlib~=1.7.4